src/requirements/myReq.txt
asgiref==3.5.1
black==22.3.0
click==8.1.3
colorama==0.4.4
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1